+++ R/generateLayerUtterances.R
@@ -32,7 +32,10 @@ generateLayerUtterances <- function(labbcat.url, match.ids, layer.id, collection
         collection_name=collection.name,
         todo="generate-now",
         utterances=paste(match.ids,collapse="\n"))
-    resp <- http.post(labbcat.url, "generateLayerUtterances", parameters)
+    resp <- http.post(labbcat.url, "edit/generateLayerUtterances", parameters)
+    if (httr::status_code(resp) != 404) { # fall back to old endpoint
+        resp <- http.post(labbcat.url, "generateLayerUtterances", parameters)
+    }
     
     ## check the response
     if (is.null(resp)) return()
@@ -78,7 +81,7 @@ generateLayerUtterances <- function(labbcat.url, match.ids, layer.id, collection
     }
 
     ## free the thread so it's not using server resources
-    http.get(labbcat.url, "threads", list(threadId=threadId, command="release"))
+    thread.release(labbcat.url, threadId)
 
     if (!is.null(pb)) { ## if there was a progress bar, 
         close(pb)