+++ R/updateTranscript.R
@@ -79,7 +79,7 @@ updateTranscript <- function(labbcat.url, transcript.path, no.progress=FALSE) {
     }
     
     ## free the upload thread so it's not using server resources
-    http.get(labbcat.url, "threads", list(threadId=threadId, command="release"))
+    thread.release(labbcat.url, threadId)
     
     return(transcript.id)
 }