+++ NEWS.md
@@ -1,3 +1,57 @@
+# nzilbb.labbcat 1.5-0
+
+Minimum LaBB-CAT version *20250430.1502* (preferred: *20250819.1454*)
+
+## New functions:
+
+Functions have been added to facilitate tidyverse-style piping expressions to build up a
+dataframe of search results with annotations, acoustic measurements, etc.
+
+The following functions add columns to a dataframe of results returned by `getMatches`:
+
+- *appendLabels* - pipeable version of `getMatchLabels`
+- *appendOffsets* - pipeable version of `getMatchAnnotations`
+- *appendFromPraat* - pipeable version of `processWithPraat`
+- *fragmentLabels* - pipeable version of `getFragmentAnnotations`
+
+The idea is you can chain together data extraction functions like:
+
+```
+matches <- getMatches(url, "(dis|mis).*") |>
+  appendLabels(c("phonemes", "morphology", "participant_gender")) |> # layers/meta-data
+  appendLabels("orthography", target.offset = -1) |> # previous word
+  appendOffsets("segment", annotations.per.layer = 3) |> # first three phone alignments
+  appendFromPraat( # F1/F2 of the vowel (the second segment):
+    segment.2.start, segment.2.end,
+    window.offset = 0.025,
+    praat.script = praatScriptFormants(formants = c(1,2)))
+```
+
+The following functions produce data/media files defined by fragment start/end times,
+and do not add columns to the given dataframe, but rather return a list of file names:
+
+- *fragmentTranscripts* - pipeable version of `getFragments`
+- *fragmentAudio* - pipeable version of `getSoundFragments`
+- *fragmentData* - pipeable version of `getFragmentAnnotationData`
+
+For example:
+
+```
+vivid.tokens <- getMatches(labbcat.url, "vivid")
+vivid.textgrids <- vivid.tokens |> fragmentTranscripts(c("word", "segment"))
+vivid.wavs <- vivid.tokens |> fragmentAudio()
+vivid.faces <- vivid.tokens |> 
+    fragmentData( ## png images of faces detected by mediapipe during the word
+        "mediapipe", path = "png",
+        start.column=Target.word.start, end.column=Target.word.end)
+```
+
+## Enhancements
+
+- *labbcatCredentials* and all other functions now support the alternative 'Form' HTTP
+   authorization that may be configured on the server, in addition to the usual 'Basic' 
+   authorization.
+
 # nzilbb.labbcat 1.4-0
 
 Minimum LaBB-CAT version *20250430.1502*